+++ CHANGELOG.md
@@ -55,6 +55,7 @@ All notable changes to this project will be documented in this file.
 -   Two bugs with location specific options not properly loading/saving
 -   Scrollbar on bottom of page in firefox when location bar does not fit the screen
 -   Players can no longer remove or add floors
+-   Players can no longer move their shape to other layers/floors/locations themselves
 
 ## [0.20.1] - 2020-05-11